requirements.txt
asgiref
Django
djangorestframework
djangorestframework_simplejwt
gunicorn
packaging
pillow
PyJWT
sqlparse
typing_extensions
